refactor: PerformanceGenre 엔티티 association 패키지로 이동
