+++ src/main/java/com/prography/lighton/performance/common/domain/entity/association/PerformanceGenre.java
@@ -1,7 +1,8 @@
-package com.prography.lighton.performance.common.domain.entity;
+package com.prography.lighton.performance.common.domain.entity.association;
 
 import com.prography.lighton.common.domain.BaseEntity;
 import com.prography.lighton.genre.domain.entity.Genre;
+import com.prography.lighton.performance.common.domain.entity.Performance;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.ManyToOne;
