+++ src/main/java/com/prography/lighton/performance/users/presentation/dto/response/GetPerformanceMapListResponseDTO.java
@@ -4,7 +4,7 @@ import static com.prography.lighton.performance.users.presentation.util.Performa
 
 import com.prography.lighton.genre.domain.entity.Genre;
 import com.prography.lighton.performance.common.domain.entity.Performance;
-import com.prography.lighton.performance.common.domain.entity.PerformanceGenre;
+import com.prography.lighton.performance.common.domain.entity.association.PerformanceGenre;
 import java.time.LocalDate;
 import java.util.List;
